refactor displayquiz directory
