+++ src/use_case/quiz/display_quiz/DisplayQuizInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.quiz;
+package use_case.quiz.display_quiz;
 
 public interface DisplayQuizInputBoundary {
     void execute();
