+++ src/use_case/quiz/display_quiz/DisplayQuizInputData.java
@@ -1,4 +1,4 @@
-package use_case.quiz;
+package use_case.quiz.display_quiz;
 
 public class DisplayQuizInputData{
 //    private String title;
