+++ src/use_case/quiz/display_quiz/DisplayQuizOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.quiz;
+package use_case.quiz.display_quiz;
 
 public interface DisplayQuizOutputBoundary {
     void prepareSuccessView(DisplayQuizOutputData quiz);
